vaccination feature module
