--- app/src/main/java/com/latticeonfhir/android/data/local/enums/VaccineErrorTypeEnum.kt
@@ -1,6 +1,0 @@
-package com.latticeonfhir.android.data.local.enums
-
-enum class VaccineErrorTypeEnum(val errorType: String) {
-    TIME("time"),
-    DOSE("dose");
-}
