+++ app/src/main/java/com/latticeonfhir/android/navigation/Screen.kt
@@ -33,9 +33,4 @@ sealed class Screen(val route: String) {
     data object DrugDispenseScreen : Screen("drugs_dispense")
     data object DispensePrescriptionScreen : Screen("dispense_prescription")
     data object OTCScreen : Screen("otc_screen")
-
-    data object VaccinationScreen : Screen("vaccination_screen")
-    data object AddVaccinationScreen : Screen("add_vaccination_screen")
-    data object ViewVaccinationScreen : Screen("view_vaccination_screen")
-    data object VaccinationErrorScreen : Screen("vaccination_error_screen")
 }
