--- core/model/src/main/java/com/latticeonfhir/core/model/enums/VaccineErrorTypeEnum.kt
@@ -1,6 +1,0 @@
-package com.latticeonfhir.core.model.enums
-
-enum class VaccineErrorTypeEnum(val errorType: String) {
-    TIME("time"),
-    DOSE("dose");
-}
