+++ features/vaccination/build.gradle.kts
@@ -4,7 +4,7 @@ plugins {
 }
 
 android {
-    namespace = "com.latticeonfhir.android.utils"
+    namespace = "com.latticeonfhir.android.vaccination"
     compileSdk = 35
 
     defaultConfig {
@@ -33,12 +33,30 @@ android {
 }
 
 dependencies {
+    implementation(projects.core.base)
+    implementation(projects.core.utils)
+    implementation(projects.core.ui)
+    implementation(projects.core.model)
+    implementation(projects.theme)
 
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
     implementation(libs.material)
+    implementation(libs.androidx.foundation.android)
+    implementation(libs.dagger.hilt.android)
+    implementation(libs.androidx.material3.android)
+    implementation(libs.androidx.navigation.runtime.ktx)
+    implementation(libs.androidx.hilt.navigation.compose)
+    implementation(libs.androidx.compose.material.iconsExtended)
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.test.espresso.core)
+    implementation(libs.okhttp.logging)
+    implementation(libs.androidx.camera.core)
+    implementation(libs.androidx.camera.camera2)
+    implementation(libs.androidx.camera.lifecycle)
+    implementation(libs.androidx.camera.view)
+    implementation(libs.androidx.camera.extensions)
+    implementation(libs.coil.kt.compose)
     implementation(libs.timber)
 }
