+++ features/vaccination/src/main/java/com/latticeonfhir/android/vaccination/data/enums/VaccineErrorTypeEnum.kt
@@ -1,0 +1,6 @@
+package com.latticeonfhir.android.vaccination.data.enums
+
+enum class VaccineErrorTypeEnum(val errorType: String) {
+    TIME("time"),
+    DOSE("dose");
+}
