+++ features/vaccination/src/main/res/drawable/delete_icon.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="48dp"
+    android:height="48dp"
+    android:tint="?attr/colorControlNormal"
+    android:viewportWidth="960"
+    android:viewportHeight="960">
+    <path
+        android:fillColor="@android:color/white"
+        android:pathData="M253,865Q212.79,865 185.39,837.9Q158,810.8 158,771L158,218L100,218L100,124L331,124L331,76L628,76L628,124L860,124L860,218L802,218L802,771Q802,810.05 774.23,837.53Q746.46,865 707,865L253,865ZM707,218L253,218L253,771Q253,771 253,771Q253,771 253,771L707,771Q707,771 707,771Q707,771 707,771L707,218ZM354,693L431,693L431,294L354,294L354,693ZM529,693L607,693L607,294L529,294L529,693ZM253,218L253,218L253,771Q253,771 253,771Q253,771 253,771L253,771Q253,771 253,771Q253,771 253,771L253,218Z" />
+</vector>
