+++ features/vaccination/src/main/res/drawable/schedule.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="18dp"
+    android:height="18dp"
+    android:viewportWidth="18"
+    android:viewportHeight="18">
+  <path
+      android:pathData="M11.75,12.916L12.917,11.749L9.834,8.666V4.833H8.167V9.333L11.75,12.916ZM9,17.333C7.848,17.333 6.764,17.114 5.75,16.676C4.736,16.239 3.854,15.645 3.104,14.895C2.354,14.145 1.761,13.263 1.323,12.249C0.886,11.236 0.667,10.152 0.667,8.999C0.667,7.847 0.886,6.763 1.323,5.749C1.761,4.735 2.354,3.854 3.104,3.104C3.854,2.354 4.736,1.76 5.75,1.322C6.764,0.885 7.848,0.666 9,0.666C10.153,0.666 11.236,0.885 12.25,1.322C13.264,1.76 14.146,2.354 14.896,3.104C15.646,3.854 16.24,4.735 16.677,5.749C17.115,6.763 17.334,7.847 17.334,8.999C17.334,10.152 17.115,11.236 16.677,12.249C16.24,13.263 15.646,14.145 14.896,14.895C14.146,15.645 13.264,16.239 12.25,16.676C11.236,17.114 10.153,17.333 9,17.333Z"
+      android:fillColor="#867D19"/>
+</vector>
