+++ features/vaccination/src/main/res/drawable/syringe.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <group>
+    <clip-path
+        android:pathData="M0,0h24v24h-24z"/>
+    <path
+        android:pathData="M3.9,11.175C3.717,10.975 3.625,10.738 3.625,10.463C3.625,10.188 3.717,9.95 3.9,9.75L6.7,6.95L5.625,5.875L5.325,6.175C5.125,6.375 4.887,6.475 4.613,6.475C4.338,6.475 4.1,6.375 3.9,6.175C3.717,5.992 3.625,5.759 3.625,5.475C3.625,5.192 3.717,4.959 3.9,4.775L5.9,2.775C6.1,2.575 6.338,2.475 6.613,2.475C6.887,2.475 7.125,2.575 7.325,2.775C7.508,2.959 7.6,3.192 7.6,3.475C7.6,3.759 7.508,3.992 7.325,4.175L7.025,4.475L8.1,5.55L10.9,2.75C11.1,2.55 11.337,2.45 11.613,2.45C11.887,2.45 12.125,2.55 12.325,2.75C12.525,2.95 12.625,3.188 12.625,3.463C12.625,3.738 12.525,3.975 12.325,4.175L11.65,4.825L19.025,12.2C19.408,12.583 19.6,13.054 19.6,13.613C19.6,14.171 19.408,14.642 19.025,15.025L18.325,15.75L23.05,20.45H20.2L16.9,17.15L16.2,17.875C15.817,18.258 15.346,18.45 14.788,18.45C14.229,18.45 13.758,18.258 13.375,17.875L6,10.5L5.325,11.175C5.125,11.358 4.887,11.45 4.613,11.45C4.338,11.45 4.1,11.358 3.9,11.175ZM7.4,9.1L14.775,16.475L17.6,13.625L16.1,12.1L14.7,13.5C14.5,13.684 14.262,13.779 13.988,13.788C13.712,13.796 13.483,13.708 13.3,13.525C13.1,13.325 13,13.088 13,12.813C13,12.538 13.1,12.3 13.3,12.1L14.7,10.7L13.2,9.2L11.8,10.6C11.6,10.783 11.363,10.875 11.087,10.875C10.813,10.875 10.575,10.783 10.375,10.6C10.192,10.4 10.1,10.163 10.1,9.888C10.1,9.613 10.192,9.375 10.375,9.175L11.775,7.775L10.25,6.25L7.4,9.1Z"
+        android:fillColor="#65558F"/>
+  </group>
+</vector>
