+++ features/vaccination/src/main/res/values/strings.xml
@@ -1,0 +1,65 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="vaccination">Vaccination</string>
+    <string name="vaccination_info">%d upcoming, %d missed, %d taken</string>
+    <string name="age_string">Age: %s</string>
+    <string name="at_birth">At birth</string>
+    <string name="completed_vaccine_info">(%d/%d completed)</string>
+    <string name="taken_on_date">Taken on %s</string>
+    <string name="due_on_date">Due on %s</string>
+    <string name="no_missed_vaccines">No Missed Vaccines</string>
+    <string name="all_vaccines_on_schedule">All vaccines are on schedule.</string>
+    <string name="overdue_count">Overdue (%d)</string>
+    <string name="no_vaccines_recorded">No Vaccines Recorded</string>
+    <string name="vaccine_log_info">Once you log a vaccination, it will appear here</string>
+    <string name="add_vaccination">Add vaccination</string>
+    <string name="status_info">Status:</string>
+    <string name="upcoming_on_info">Upcoming on %s</string>
+    <string name="date">Date:</string>
+    <string name="number_dose">%s Dose</string>
+    <string name="search_vaccination">Search vaccination</string>
+    <string name="lot_no_mandatory">"Lot number * "</string>
+    <string name="date_of_expiry_mandatory">Date of expiry *</string>
+    <string name="date_format">DD/MM/YYYY</string>
+    <string name="vaccine_manufacturer">Vaccine manufacturer</string>
+    <string name="notes_heading">Notes</string>
+    <string name="notes_for_adverse_reaction">Enter the adverse reactions if any</string>
+    <string name="upload_certifications">Upload certifications</string>
+    <string name="upload_certifications_info">Upload up to 10 files, each of size 5 MB or less.\nFile types: jpg/jpeg</string>
+    <string name="upload">Upload</string>
+    <string name="upload_from_files">Upload from files</string>
+    <string name="upload_from_gallery">Upload from gallery</string>
+    <string name="take_a_picture">Take a picture</string>
+    <string name="yes_delete">Yes, delete</string>
+    <string name="confirm_deletion">Confirm deletion</string>
+    <string name="confirm_deletion_info">Are you sure you want to delete this</string>
+    <string name="view_vaccination">View vaccination</string>
+    <string name="lot_no_label">Lot number:</string>
+    <string name="date_of_expiry_label">Date of expiry:</string>
+    <string name="vaccine_manufacturer_label">Vaccine manufacturer:</string>
+    <string name="uploaded_certificates">Uploaded certificates</string>
+    <string name="vaccine_time_window_error_message">Vaccine entry is allowed only during the designated time window.</string>
+    <string name="vaccine_time_window_error_message_info">Please record details during that period.</string>
+    <string name="vaccine_previous_dose_error_message_info">Vaccine sequence must be followed.</string>
+    <string name="vaccine_previous_dose_error_message">Please record the previous dose(s) before entering this dose.</string>
+    <string name="vaccine_name_with_code">%s (%s)</string>
+    <string name="fully_vaccinated">Fully vaccinated</string>
+    <string name="vaccination_added_successfully">Vaccination added successfully!</string>
+    <string name="vaccine_view_name_info">%s(%s) - %s Dose - %s</string>
+    <string name="vaccine_date_range">(%s - %s)</string>
+    <string name="confirm_dob_update">Confirm D.O.B. update</string>
+    <string name="confirm_dob_update_description">If you change your DOB, only future vaccine dates will update; taken doses remain unchanged. Do you wish to proceed?</string>
+    <string name="yes_continue">Yes, continue</string>
+    <string name="fetching_immunization_recommendations">Fetching immunization recommendations</string>
+    <string name="fetching_immunization_recommendations_info">Immunization guidelines are stored on a central server. Please give us a few minutes to fetch them. Once they have been retrieved, you will be able to work offline.</string>
+    <string name="please_grant_permission">Please grant permission.</string>
+    <string name="save">Save</string>
+    <string name="no_results_found">No results found</string>
+    <string name="no_go_back">No, go back</string>
+    <string name="permissions_required">Permission required</string>
+    <string name="permissions_required_description">To continue, this app needs access to your camera and storage. Please allow these permissions from the settings app.</string>
+    <string name="go_to_settings">Go to settings</string>
+    <string name="cancel">Cancel</string>
+    <string name="dismiss">Dismiss</string>
+    <string name="notes_colon">Notes:</string>
+</resources>
